# openapi.yaml
openapi: 3.0.0
info:
  title: Мессенджер API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Основной сервер
paths:
...
  /auth/authorize:
    get:
      tags:
        - Auth
      summary: Получение данных своей учетнйо записи по access токену
      operationId: authAuthorizeGET
      responses:
        '200':
          description: Данные успешно получены
        '401':
          description: User ID не найден в токене
      x-swagger-router-controller: Auth
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Авторизация пользователя
      operationId: authLoginPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth_login_body'
        required: true
      responses:
        '200':
          description: Успешный вход
        '401':
          description: Неверные учетные данные
        '404':
          description: Пользователь не найден
      x-swagger-router-controller: Auth
...
  /auth/google-login:
    get:
      tags:
        - Auth
      summary: Начало процесса входа через Google
      operationId: authGoogleLoginGET
      responses:
        '302':
          description: Перенаправление на страницу входа Google
        '400':
          description: Ошибка при начале процесса входа через Google
      x-swagger-router-controller: Auth
  /auth/google-response:
    get:
      tags:
        - Auth
      summary: Обработка ответа от Google после входа
      operationId: authGoogleResponseGET
      responses:
        '200':
          description: Успешный вход или регистрация через Google
        '400':
          description: Ошибка аутентификации через Google
        '401':
          description: Неверные учетные данные Google
      x-swagger-router-controller: Auth
...
components:
  schemas:
    auth_login_body:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...